requirements.txt
Django==2.1.5
django-bootstrap4==0.0.7
django-chartjs==1.3
pytz==2018.9
six==1.12.0
matplotlib==3.0.2
networkx==2.2
django-pyodbc-azure==2.1.0.0
pyodbc==4.0.25
requests==2.21.0